+++ src/main/java/com/tesa/hospitalerd/service/interfaces/EquipmentService.java
@@ -1,7 +1,9 @@
 package com.tesa.hospitalerd.service.interfaces;
 
 import com.tesa.hospitalerd.model.request.EquipmentRequest;
+import com.tesa.hospitalerd.model.request.PatientRequest;
 import com.tesa.hospitalerd.model.entity.Equipment;
+import com.tesa.hospitalerd.model.entity.Patient;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -11,11 +13,20 @@ public interface EquipmentService {
 
     void createEquipment(EquipmentRequest request);
     List<Equipment> getAllEquipments();
-    Equipment getEquipmentById(int id);
+    Equipment getEquipmentById(Long id);
     List<Equipment> searchEquipment(String query);
     List<Equipment> getEquipmentByStatus(String status);
-    void markEquipmentForMaintenance(int id);
-    void markEquipmentAsAvailable(int id);
-    void decommissionEquipment(int id);
-    void updateEquipment(int id, EquipmentRequest request);
+    void markEquipmentForMaintenance(Long id);
+    void markEquipmentAsAvailable(Long id);
+    void decommissionEquipment(Long id);
+    void updateEquipment(Long id, EquipmentRequest request);
+
+//    interface PatientService {
+//
+//        void createPatient(PatientRequest request);
+//        List<Patient> getAllPatients();
+//        Patient getPatientById(int id);
+//        List<Patient> searchPatient(String query);
+//        void updatePatient(int id, PatientRequest request);
+//    }
 }
